JWT working nicely, expires the token every XXX seconds, once expired the razilo components request element will resend for a refreshed token, if successfull then it will also re send the original request and fullfill it.
This offers seemless handshaking of JWT's away from your application. If a user has not refreshed in XXX seconds, then the refresh will not be allowed (idle time of a token that prevents refresh on long idle tokens)

Now we can log in, and refresh tokens, to log out just clear local storage of token in UI. Now we can look to tidy up a bit and start rebuilding the UI in raziloComponent giving us a way more secure application.